+++ ui/contractVerification/fields/ContractVerificationFieldContractIndex.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { useFormContext } from 'react-hook-form';
 
 import type { FormFields } from '../types';
-import type { Option } from 'ui/shared/FancySelect/types';
+import type { Option } from 'ui/shared/forms/inputs/select/types';
 
 import FormFieldFancySelect from 'ui/shared/forms/fields/FormFieldFancySelect';
