+++ ui/shared/forms/inputs/select/__screenshots__/FancySelect.pw.tsx_dark-color-mode_size-md-dark-mode-filled-1.png
