+++ ui/shared/forms/inputs/select/__screenshots__/FancySelect.pw.tsx_default_size-lg-dark-mode-read-only-1.png
