+++ ui/shared/forms/inputs/select/__screenshots__/FancySelect.pw.tsx_default_size-md-dark-mode-error-1.png
